project_root/requirements.txt
python-telegram-bot[job-queue]==20.8
openai==1.12.0
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
SQLAlchemy==2.0.27
aiosqlite==0.20.0
aiohttp==3.9.3
Pillow==10.1.0